[feat] 임시 컴포넌트 테스트 뷰 컨트롤러 생성
